requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.56.0
anyio==4.9.0
attrs==25.3.0
authlib==1.6.0
backoff==2.2.1
beautifulsoup4==4.13.4
boto3==1.38.28
botocore==1.38.28
browser-use==0.4.2
bubus==1.3.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
cython==3.1.2
defusedxml==0.7.1
distro==1.9.0
faiss-cpu==1.11.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-genai==1.24.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.29.0
grpcio==1.73.1
grpcio-status==1.72.1
h11==0.16.0
h2==4.2.0
hf-xet==1.1.2
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.3
hyperframe==6.1.0
idna==3.10
jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.22
langchain-anthropic==0.3.3
langchain-aws==0.2.19
langchain-core==0.3.49
langchain-deepseek==0.1.3
langchain-google-genai==2.1.2
langchain-ollama==0.3.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langsmith==0.3.44
markdownify==1.1.0
markupsafe==3.0.2
mem0ai==0.1.113
monotonic==1.6
mpmath==1.3.0
multidict==6.6.2
networkx==3.5
numpy==2.3.1
ollama==0.5.1
openai==1.93.0
orjson==3.10.18
packaging==24.2
patchright==1.52.5
pillow==11.2.1
playwright==1.53.0
portalocker==2.10.1
posthog==6.0.2
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pyee==13.0.0
pyjwt==2.10.1
pyobjc-core==11.1
pyobjc-framework-cocoa==11.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
qdrant-client==1.14.3
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
ruff==0.12.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
screeninfo==0.8.1
sentence-transformers==4.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.4
twilio==9.6.3
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.5.0
uuid7==0.1.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.23.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.1
